Restructure the .restart file reader, and read the rest of the header
Now resolves https://github.com/jftsang/mdpmflc/issues/12, and is ready
to be extended to read other parts of the .restart file.
I think it will be better in the long run to read the entire .restart
file, rather than the .data and .fstat files. It isn't necessarily any
slower, either, since we can seek based on the number of lines.