First step towards making animations for .data files.
Part of <#7>. The current strategy is to produce .gif files locally and
then convert them to .webm and then serve the .webm's. There is
currently a bug with matplotlib that mp4 cannot be created:
<https://github.com/matplotlib/matplotlib/issues/16965>.

This strategy is *very* slow. As with the rest of the graphics (and also
with the MercuryPostprocessing library), the massive bottleneck is
reading the .data. files, which are read in their entirety even if you
want only a few thousand lines. Since the process is so long and costly,
there is the risk of submitting two requests, and the second one
overwriting the first, and duplicating the effort. Maybe use some sort
of locking.

More importantly, use queueing (<#24>).