Fix the random sampling when plotting .data. files
This change fixes a bug in which the random sample used to plot data
files is then used to calculate the frontal position, rather than using
all particles.

Also, you can now pass a sample size that is a floating-point number
less than 1, which will be interpreted as the fraction of particles that
should be plotted.