Big reorganisation of package structure and urls
The Flask app is now in app.py and URLs are registered there rather than
as decorators on each of the view (controller) functions. This sorts out
messy imports, and centralises where URLs can be found.

Test plan: Tested with some basic URLs, but didn't have any real data on
hand so I couldn't test the URLs for results or plots.